// src/components/compounds/blog/instant-gram/types/features.ts
export type Features = {
  countries?: string[];
  cities?: string[];
  districts?: string[];
  attractions?: string[];
  accommodation?: string[];
  supermarkets?: string[];
  shops?: string[];
  consumables?: string[];
  cafes?: string[];
  bakeries?: string[];
  gelaterias?: string[];
  restaurants?: string[];
  bars?: string[];
  nostalgiaEffect?: string[];
};

// For my convenience:

// "features": {
//   "countries": [],
//   "cities": [],
//   "districts": [],
//   "attractions": [],
//   "accommodation": [],
//   "supermarkets": [],
//   "shops": [],
//   "consumables": [],
//   "cafes": [],
//   "bakeries": [],
//   "gelaterias": [],
//   "restaurants": [],
//   "bars": [],
//   "nostalgiaEffect": []
// }